fix: fix a item_show for recycler view, add retrofit lib and behavior to download data from an API

+++ app/src/main/java/com/example/testshowapp/model/Rating.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Rating(
     @SerializedName("average")
-    val average: Int
+    val average: Double
 )
